Move server and webpack files to separate folder
