+++ package.json
@@ -4,11 +4,11 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "build": "webpack --config webpack.prod.babel.js",
-    "prod": "npm run build && NODE_ENV=production node_modules/.bin/babel-node server",
-    "dev": "NODE_ENV=development node_modules/.bin/babel-node server.js",
+    "build": "webpack --config webpack/webpack.prod.babel.js",
+    "prod": "npm run build && NODE_ENV=production node_modules/.bin/babel-node server/server.prod.js",
+    "dev": "NODE_ENV=development node_modules/.bin/babel-node server/server.dev.js",
     "lint": "eslint . --fix",
-    "postinstall": "webpack --config webpack.dll.babel.js"
+    "postinstall": "webpack --config webpack/webpack.dll.babel.js"
   },
   "dependencies": {
     "express": "^4.15.4",
@@ -41,7 +41,6 @@
     "path": "^0.12.7",
     "uglifyjs-webpack-plugin": "^0.4.6",
     "webpack": "^3.5.6",
-    "webpack-bundle-analyzer": "^2.9.0",
     "webpack-dev-middleware": "^1.12.0",
     "webpack-dev-server": "^2.7.1",
     "webpack-hot-middleware": "^2.19.1",
