+++ server/server.prod.js
@@ -1,0 +1,20 @@
+import path from 'path'
+import express from 'express'
+
+
+const PORT = process.env.PORT || 8080
+const DIST_DIR = path.resolve(__dirname, '..', 'public')
+
+const app = express()
+
+app.use(express.static(DIST_DIR))
+app.get('*', (req, res) => {
+  res.sendFile(path.resolve(__dirname, 'public', '..', 'index.html'))
+})
+
+app.listen(PORT, (error) => {
+  if (error) {
+    console.log(error)
+  }
+  console.info('Express is listening on PORT %s.', PORT)
+})
