+++ webpack/webpack.common.js
@@ -1,12 +1,11 @@
 import webpack from 'webpack'
 import HtmlWebpackPlugin from 'html-webpack-plugin'
 import CleanWebpackPlugin from 'clean-webpack-plugin'
-import WebpackBundleAnalyzer from 'webpack-bundle-analyzer'
 import CopyWebpackPlugin from 'copy-webpack-plugin'
 import path from 'path'
 
+import dllManifest from '../public/vendor/vendor-manifest.json'
 
-const BundleAnalyzerPlugin = WebpackBundleAnalyzer.BundleAnalyzerPlugin
 
 const DIST_DIR = 'public'
 const SRC_DIR = 'src'
@@ -15,7 +14,7 @@ export default {
   output: {
     publicPath: '/',
     filename: '[name]-[hash].min.js',
-    path: path.resolve(__dirname, DIST_DIR),
+    path: path.resolve(__dirname, '..', DIST_DIR),
   },
   plugins: [
     new CleanWebpackPlugin(
@@ -26,23 +25,18 @@ export default {
     ),
     new webpack.DllReferencePlugin({
       context: '.',
-      manifest: require('./public/vendor/vendor-manifest.json'),
+      manifest: dllManifest,
     }),
     new HtmlWebpackPlugin({
       template: 'src/_tpl/index.html',
       inject: 'body',
       filename: 'index.html',
     }),
-    new BundleAnalyzerPlugin({
-      analyzerMode: 'static',
-      openAnalyzer: false,
-      reportFilename: 'report.html',
-    }),
     new CopyWebpackPlugin([
-      { from: path.join(SRC_DIR, '/_tpl/favicon.ico'), to: 'favicon.ico' },
-      { from: path.join(SRC_DIR, '/_tpl/icon-192x192.png'), to: 'icon-192x192.png' },
-      { from: path.join(SRC_DIR, '/_tpl/icon-512x512.png'), to: 'icon-512x512.png' },
-      { from: path.join(SRC_DIR, '/_tpl/manifest.json'), to: 'manifest.json' },
+      { from: path.resolve(__dirname, '..', SRC_DIR, '_tpl', 'favicon.ico'), to: 'favicon.ico' },
+      { from: path.resolve(__dirname, '..', SRC_DIR, '_tpl', 'icon-192x192.png'), to: 'icon-192x192.png' },
+      { from: path.resolve(__dirname, '..', SRC_DIR, '_tpl', 'icon-512x512.png'), to: 'icon-512x512.png' },
+      { from: path.resolve(__dirname, '..', SRC_DIR, '_tpl', 'manifest.json'), to: 'manifest.json' },
     ]),
   ],
   module: {
